test : 특정 게시물 정보 조회 테스트 작성 및 PostInfoResponse 수정 #27
